+++ src/main/java/com/kaust/ms/manager/prompt/chat/infrastructure/ia/model/BiomedicalRequest.java
@@ -14,6 +14,6 @@ public class BiomedicalRequest {
     /** query to be executed. */
     private String query;
     /** temperature in Celsius. */
-    private Integer temperature;
+    private Double temperature;
 
 }
