+++ src/main/java/com/kaust/ms/manager/prompt/settings/domain/models/response/ModelGlobalResponse.java
@@ -16,6 +16,6 @@ public class ModelGlobalResponse {
     /** model. */
     private String model;
     /** quantityCreativity. */
-    private Integer quantityCreativity;
+    private Double quantityCreativity;
 
 }
